webpage/voice/pyproject.toml
[project]
name = "Script-Voice"
version = "0.1.0"
description = "Create the voice for the video"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hume[microphone]>=0.12.1",
    "python-dotenv>=1.1.1",
]
